refactor common asarray
